IVrm10SpringBoneRuntime.InitializeAsync
